test(bench): isolate rig lease coverage
